GestureRecorder: use uint16_t, so we can store double the amount of frames

+++ GlissModes.cpp
@@ -1174,7 +1174,7 @@ public:
 
 #ifdef ENABLE_RECORDER_MODE
 static constexpr size_t kNumRecs = 4;
-typedef float RecorderSampleT;
+typedef uint16_t RecorderSampleT;
 static std::array<RecorderSampleT,kMaxRecordBytes / sizeof(RecorderSampleT)> recorderData;
 class GestureRecorder
 {
